stream based buffers implemented in link

+++ Header.cs
@@ -8,7 +8,7 @@ namespace Lynx
         [JsonPropertyName("T")] public MessageType Type { get; }
         [JsonPropertyName("H")] public string Handler { get; }
         [JsonPropertyName("C")] public string Command { get; }
-        [JsonPropertyName("S")] public int ContentSize { get; set; }
+        [JsonPropertyName("S")] public long ContentSize { get; set; }
 
         public Header(long id, MessageType type, string handler, string command)
         {
